+++ web/src/components/Calendar/RecurrenceModal.tsx
@@ -3,7 +3,6 @@ import {
   Modal,
   Form,
   Select,
-  Input,
   DatePicker,
   Checkbox,
   Space,